requirements.txt
asgiref==3.8.1
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
